* develop:
  feat: [C4R-716] add validation in env.validation
  fix: fixed lambda function (wrong http code)
  fix: correctly cast boolean values coming from form data for organization requests
  refactor: updated script
  refactor: updated echo texts for clarity
  fix: admin remove application from organization failure
  fix: update user fails on unique phone number on own entry
  feat: added a shell script to backup remote db and restore into local db
  feat(organization): update completion status on update of financial, report, investor, partner data
  refactor(organization): update organization filters and table to support new completionStatus column
  refactor(organization-view): update completion status

+++ backend/src/common/base/base-dao.class.ts
@@ -76,13 +76,7 @@ export abstract class BaseDAO<T> {
   async remove(findCriteria: FindManyOptions<T>) {
     const records = await this.getMany(findCriteria);
 
-    if (records.length === 0) {
-      throw new Error(
-        `Could not find the records by ${JSON.stringify(findCriteria)}`,
-      );
-    }
-
-    return this.repository.remove(records);
+    if (records.length > 0) return this.repository.remove(records);
   }
 
   count(findCriteria?: FindManyOptions<T>): Promise<number> {
